Fixed flag size from uint8_t to uint16_t
We ran out of flags (we have 9 now) so the size of the flag
is not sufficient.
In theory there should not be any problem but it is not clear
if this flag is used to be written somewhere where it is expected to
be 8 bits.

To be checked later

+++ drivers/block-vhd.c
@@ -143,7 +143,7 @@ unsigned int SPB;
 #define VHD_FLAG_TX_LIVE             1
 #define VHD_FLAG_TX_UPDATE_BAT       2
 
-typedef uint8_t vhd_flag_t;
+typedef uint16_t vhd_flag_t;
 
 struct vhd_state;
 struct vhd_request;
